Fix mem leak and fix some compiler warnings
This fixes a memory leak in Record identified by scan-build, and fixes some assignment
issues that Clang was warning about.

+++ src/containers/MerkleTree.cpp
@@ -217,8 +217,9 @@ bool MerkleTree::verifyPath(const Json::Value& path, const RecordPtr& record)
   if (path[0]["hash"] != leaf.getBase64Hash())
     return false;
 
-  for (int j = 1; j < path.size(); j++)
+  for (size_t j = 1; j < path.size(); j++)
   {
+    //todo:
     // if previous hashes do not match either current match, return false
     // concatenateHashes
     // trouble is that I need the neighboring Record, yes? no
